requirements.txt
websockets>=12.0
python-dotenv
